requirements.txt
Flask
requests
elevenlabs
python-dotenv